+++ packages/webview_flutter/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.3.2+1
+
+* Log a more detailed warning at build time about the previous AndroidX
+  migration.
+
+## 0.3.2
+
+* Added CookieManager to interface with WebView cookies. Currently has the ability to clear cookies.
+
 ## 0.3.1
 
 * Added JavaScript channels to facilitate message passing from JavaScript code running inside